+++ completions/zsh/_swaymsg
@@ -23,6 +23,7 @@ types=(
 'get_bar_config'
 'get_version'
 'get_binding_modes'
+'get_bindings'
 'get_config'
 'send_tick'
 )
