+++ include/sway/config.h
@@ -354,9 +354,10 @@ enum ipc_feature {
 	IPC_FEATURE_EVENT_BINDING = 4096,
 	IPC_FEATURE_EVENT_INPUT = 8192,
 	IPC_FEATURE_GET_SEATS = 16384,
+	IPC_FEATURE_GET_BINDINGS = 32768,
 
 	IPC_FEATURE_ALL_COMMANDS =
-		1 | 2 | 4 | 8 | 16 | 32 | 64 | 128 | 16384,
+		1 | 2 | 4 | 8 | 16 | 32 | 64 | 128 | 16384 | 32768,
 	IPC_FEATURE_ALL_EVENTS = 256 | 512 | 1024 | 2048 | 4096 | 8192,
 
 	IPC_FEATURE_ALL = IPC_FEATURE_ALL_COMMANDS | IPC_FEATURE_ALL_EVENTS,
