+++ swaymsg/main.c
@@ -417,6 +417,8 @@ int main(int argc, char **argv) {
 		type = IPC_GET_VERSION;
 	} else if (strcasecmp(cmdtype, "get_binding_modes") == 0) {
 		type = IPC_GET_BINDING_MODES;
+	} else if (strcasecmp(cmdtype, "get_bindings") == 0) {
+		type = IPC_GET_BINDINGS;
 	} else if (strcasecmp(cmdtype, "get_config") == 0) {
 		type = IPC_GET_CONFIG;
 	} else if (strcasecmp(cmdtype, "send_tick") == 0) {
